fix: support multiple domains
Certbot supports registering multiple domains under the root domain
from the command line. The Joker API, however, requires to update all
domain entries in the same API call.

This patch stores the domains in an array to send them together
for each call from certbot in a session.

Note that it doesn't retain the TXT entries set for the root domain
before certbot was executed or after.

Signed-off-by: Chris Zankel <chris@zankel.net>